+++ python/tests/kat/t_headerrouting.py
@@ -27,6 +27,7 @@ class HeaderRoutingTest(MappingTest):
 apiVersion: ambassador/v0
 kind:  Mapping
 name:  {self.name}-target1
+host: "*"
 prefix: /{self.name}/
 service: http://{self.target.path.fqdn}
 """)
@@ -35,6 +36,7 @@ service: http://{self.target.path.fqdn}
 apiVersion: ambassador/v0
 kind:  Mapping
 name:  {self.name}-target2
+host: "*"
 prefix: /{self.name}/
 service: http://{self.target2.path.fqdn}
 headers:
@@ -130,6 +132,7 @@ allowed_authorization_headers:
 apiVersion: ambassador/v0
 kind:  Mapping
 name:  {self.name}-target1
+host: "*"
 prefix: /target/
 service: http://{self.target1.path.fqdn}
 """)
@@ -138,6 +141,7 @@ service: http://{self.target1.path.fqdn}
 apiVersion: ambassador/v0
 kind:  Mapping
 name:  {self.name}-target2
+host: "*"
 prefix: /target/
 service: http://{self.target2.path.fqdn}
 headers: