+++ python/tests/kat/t_queryparameter_routing.py
@@ -15,6 +15,7 @@ class QueryParameterRoutingTest(AmbassadorTest):
 apiVersion: ambassador/v2
 kind:  Mapping
 name:  {self.name}-target1
+host: "*"
 prefix: /target/
 service: http://{self.target1.path.fqdn}
 """)
@@ -23,6 +24,7 @@ service: http://{self.target1.path.fqdn}
 apiVersion: ambassador/v2
 kind:  Mapping
 name:  {self.name}-target2
+host: "*"
 prefix: /target/
 service: http://{self.target2.path.fqdn}
 query_parameters:
@@ -49,6 +51,7 @@ class QueryParameterRoutingWithRegexTest(AmbassadorTest):
 apiVersion: ambassador/v2
 kind:  Mapping
 name:  {self.name}-target
+host: "*"
 prefix: /target/
 service: http://{self.target.path.fqdn}
 regex_query_parameters:
@@ -77,6 +80,7 @@ class QueryParameterPresentRoutingTest(AmbassadorTest):
 apiVersion: ambassador/v2
 kind:  Mapping
 name:  {self.name}-target
+host: "*"
 prefix: /target/
 service: http://{self.target.path.fqdn}
 query_parameters: